backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-redis==5.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
psycopg-binary==3.3.3
cryptography==42.0.7
pyotp==2.9.0
django-ratelimit==4.1.0
python-decouple==3.8
Pillow==10.3.0
pytest-django==4.8.0
pytest==8.2.0
factory-boy==3.3.0
